Move citation related item and new related module.
